+++ DIAYN/evaluate_agent_robosuite.py
@@ -6,7 +6,7 @@ import yaml
 
 import gymnasium as gym
 
-from DIAYN import AgentBase
+from DIAYN import AgentBase, CustomGymWrapper, make_env
 from DIAYN.utils import pad_to_dim_2
 
 
@@ -16,98 +16,6 @@ from robosuite.controllers.composite.composite_controller import WholeBody
 from robosuite.wrappers import VisualizationWrapper
 
 import robosuite as suite
-from robosuite.wrappers import GymWrapper
-
-class CustomGymWrapper(gym.ObservationWrapper):
-    def __init__(self, robosuite_env, config):
-        # Wrap robosuite with GymWrapper inside
-        gym_env = GymWrapper(robosuite_env)
-        super().__init__(gym_env)
-
-        # Figure out which state variables we care about
-        self.use_eef_state = config['observations']['use_eef_state']
-
-        # Extract dimensions of observation space
-        obs_dict = robosuite_env._get_observations()
-
-        if self.use_eef_state:
-            observation_raw = np.concatenate([
-                obs_dict['robot0_eef_pos'],
-                obs_dict['robot0_eef_quat']
-            ])
-        else:
-            observation_raw = np.concatenate([
-                obs_dict['robot0_proprio-state'],
-                obs_dict['object-state']
-            ])
-
-        # Add new dimensions to observation spaceZ
-        new_dim = observation_raw.shape[0]
-        self.observation_space = gym.spaces.Box(
-            low=-np.inf,
-            high=np.inf,
-            shape=(new_dim,),
-            dtype=np.float32
-        )
-    
-    def observation(self, obs):
-        # obs is the flat vector from GymWrapper
-        # Add x-position (or whatever custom feature you want)
-        obs_dict = 	self.env.unwrapped._get_observations()
-
-        if (self.use_eef_state):
-            # print("using end effector state")
-            # print("Using end effector state")
-            # print("eef pos: " + str(obs_dict['robot0_eef_pos']))
-            # print("eef quat: " + str(obs_dict['robot0_eef_quat']))
-            observation_raw = np.concatenate([obs_dict['robot0_eef_pos'], obs_dict['robot0_eef_quat']])
-        else:
-            observation_raw = np.concatenate([obs_dict['robot0_proprio-state'], obs_dict['object-state']])
-
-        # print("obs dict from custom gym wrapper: " + str(obs_dict))
-        # print("observation_raw: " + str(observation_raw))
-        return observation_raw
-
-def make_env(env_name, robots):
-    def _thunk():
-        print("Creating new environment")
-
-        controller_config = load_composite_controller_config(
-                controller=None,
-                robot="Panda",
-        )
-
-        robosuite_config = {
-            "env_name": env_name,
-            "robots": robots,
-            "controller_configs": controller_config,
-        }
-
-        robosuite_env = suite.make(
-            **robosuite_config,
-            has_renderer=False,
-            has_offscreen_renderer=False,
-            render_camera="agentview",
-            ignore_done=True,
-            use_camera_obs=False,
-            reward_shaping=True,
-            control_freq=20,
-            hard_reset=False,
-        )
-
-        with open("./config/ur5e_config.yaml", "r") as file:
-            config = yaml.safe_load(file)
-
-        env = CustomGymWrapper(robosuite_env, config)
-
-        # Ensure metadata exists and is a dict before modifying
-        if env.metadata is None:
-            env.metadata = {}
-        env.metadata["render_modes"] = []
-        env.metadata["autoreset"] = False
-
-        return env
-    return _thunk
 
 def evaluate_agent_robosuite(
     environment_name: str,
@@ -118,6 +26,7 @@ def evaluate_agent_robosuite(
     device,
     skill_index: Optional[int] = None,
     num_skills: Optional[int] = None,
+    config: Optional[dict] = None,
 ):
     """Evaluate Agent Interaction. Run until episode ends or max_episode_step is reached.
 
@@ -146,7 +55,7 @@ def evaluate_agent_robosuite(
             dim=-1,
         )
 
-    env = make_env(environment_name, robots)()
+    env = make_env(config)()
 
     episode_total_return = []
     episode_length = []
