+++ examples/visualize_diayn_ur5e.py
@@ -4,108 +4,14 @@ import yaml
 import numpy as np
 import torch
 import gymnasium as gym
+from pathlib import Path
 
-from DIAYN import DIAYNAgent, evaluate_agent_robosuite, visualize_robosuite
+from DIAYN import DIAYNAgent, CustomGymWrapper, make_env, evaluate_agent_robosuite, visualize_robosuite
 
 import robosuite as suite
 from robosuite import load_composite_controller_config
 
 from gymnasium.vector import SyncVectorEnv
-from robosuite.wrappers import GymWrapper
-
-
-class CustomGymWrapper(gym.ObservationWrapper):
-    def __init__(self, robosuite_env, config):
-        # Wrap robosuite with GymWrapper inside
-        gym_env = GymWrapper(robosuite_env)
-        super().__init__(gym_env)
-
-        # Figure out which state variables we care about
-        self.use_eef_state = config['observations']['use_eef_state']
-
-        # Extract dimensions of observation space
-        obs_dict = robosuite_env._get_observations()
-
-        if self.use_eef_state:
-            observation_raw = np.concatenate(
-                [obs_dict['robot0_eef_pos'], obs_dict['robot0_eef_quat']]
-            )
-        else:
-            observation_raw = np.concatenate(
-                [obs_dict['robot0_proprio-state'], obs_dict['object-state']]
-            )
-
-        # Add new dimensions to observation spaceZ
-        new_dim = observation_raw.shape[0]
-        self.observation_space = gym.spaces.Box(
-            low=-np.inf, high=np.inf, shape=(new_dim,), dtype=np.float32
-        )
-
-    def observation(self, obs):
-        # obs is the flat vector from GymWrapper
-        # Add x-position (or whatever custom feature you want)
-        obs_dict = self.env.unwrapped._get_observations()
-
-        if self.use_eef_state:
-            # print("using end effector state")
-            # print("Using end effector state")
-            # print("eef pos: " + str(obs_dict['robot0_eef_pos']))
-            # print("eef quat: " + str(obs_dict['robot0_eef_quat']))
-            observation_raw = np.concatenate(
-                [obs_dict['robot0_eef_pos'], obs_dict['robot0_eef_quat']]
-            )
-        else:
-            observation_raw = np.concatenate(
-                [obs_dict['robot0_proprio-state'], obs_dict['object-state']]
-            )
-
-        # print("obs dict from custom gym wrapper: " + str(obs_dict))
-        # print("observation_raw: " + str(observation_raw))
-        return observation_raw
-
-
-def make_env():
-    def _thunk():
-        print('Creating new environment')
-
-        controller_config = load_composite_controller_config(
-            controller=None,
-            robot='Panda',
-        )
-
-        robosuite_config = {
-            'env_name': 'Lift',
-            'robots': 'Panda',
-            'controller_configs': controller_config,
-        }
-
-        robosuite_env = suite.make(
-            **robosuite_config,
-            has_renderer=False,
-            has_offscreen_renderer=False,
-            render_camera="birdview",
-            ignore_done=True,
-            use_camera_obs=False,
-            reward_shaping=True,
-            control_freq=20,
-            hard_reset=False,
-        )
-
-        with open('./examples/config/ur5e_config.yaml', 'r') as file:
-            config = yaml.safe_load(file)
-
-        env = CustomGymWrapper(robosuite_env, config)
-
-        # Ensure metadata exists and is a dict before modifying
-        if env.metadata is None:
-            env.metadata = {}
-        env.metadata['render_modes'] = []
-        env.metadata['autoreset'] = False
-
-        return env
-
-    return _thunk
-
 
 def main(
     environment_name: str,
@@ -127,7 +33,7 @@ def main(
     # Setup env variables
 
     # TODO: Just get dims directly instead of dealing with this each time
-    envs = SyncVectorEnv([make_env() for _ in range(num_envs)])
+    envs = SyncVectorEnv([make_env(config) for _ in range(num_envs)])
 
     observation_dims = envs.observation_space.shape[1]
     action_dims = envs.action_space.shape[1]
@@ -221,6 +127,7 @@ def main(
                 device,
                 skill_index=z,
                 num_skills=num_skills,
+                config= config,
             )
             skills_return.append(np.mean(result_dict['total_return']))
             skills_mean_step_reward.append(
@@ -285,6 +192,7 @@ def main(
             device,
             skill_index=visualize_skill,
             num_skills=num_skills,
+            config=config,
         )
         print(
             f'[Skill {visualize_skill}] Mean Step Reward: {np.nanmean(result_dict["total_return"]/result_dict["episode_length"])} Mean Total Return: {np.mean(result_dict["total_return"])}'
@@ -304,9 +212,12 @@ def checkpoint_check(filepath: str):
 
 if __name__ == '__main__':
     # Read config file for all settings
-    with open('./examples/config/ur5e_config.yaml', 'r') as file:
-        config = yaml.safe_load(file)
+    current_dir = str(Path(__file__).parent.resolve())
+    print('Current directory:', current_dir)
 
+    with open(current_dir + '/config/ur5e_config.yaml', 'r') as file:
+        config = yaml.safe_load(file)
+    # exit()
     environment_name = config['params']['environment_name']
     robots = config['params']['robots']
     num_envs = config['params']['num_envs']
